Move models from Coins api entry to models dedicqted package
